# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/positions": {
            "post": {
                "description": "Create a new position with the input data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "positions"
                ],
                "summary": "Create a new position",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/positions/{id}": {
            "get": {
                "description": "Get details of the position",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "positions"
                ],
                "summary": "Get details of the position",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Position ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pages.Position"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Response": {
            "type": "object",
            "properties": {
                "Status": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "pages.Position": {
            "type": "object",
            "properties": {
                "current_price": {
                    "type": "number"
                },
                "date_created": {
                    "type": "string"
                },
                "direction": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "instrument": {
                    "type": "string"
                },
                "margin": {
                    "type": "number"
                },
                "price": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                },
                "result": {
                    "type": "number"
                },
                "stop_loss": {
                    "type": "string"
                },
                "take_profit": {
                    "type": "string"
                },
                "trailing_stop": {
                    "type": "string"
                }
            }
        }
    }
}